Fix problems with warriors loading
When trying to load the same warrior twice in a row the input control
doesn't emit the onload event. Because of this it was necessary to load
the warrior, try to load it again but cancel the window, and then load
it.

Now the input control is replaced with a clone of itself after loading a
warrior so this bug is fixed.